myapp-backend/requirements.txt
alembic==1.14.1
asttokens==3.0.0
backcall==0.2.0
bidict==0.23.1
blinker==1.8.2
cffi==1.17.1
click==8.1.8
comm==0.2.2
cryptography==3.3
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
flask==3.0.3
Flask-Cors==5.0.0
Flask-Login==0.6.3
flask-mail==0.10.0
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.16.0
importlib-metadata==8.5.0
importlib-resources==6.4.5
ipykernel==6.29.5
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.6
jupyter-client==8.6.3
jupyter-core==5.8.1
mako==1.3.10
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
msgpack-python==0.4.8
mysql-connector-python==9.0.0
nest-asyncio==1.6.0
numpy==1.24.4
packaging==25.0
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.3.6
prompt-toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pycparser==2.22
pygments==2.19.1
python-dateutil==2.9.0.post0
python-engineio==4.12.3
python-socketio==5.13.0
pytz==2025.2
pyzmq==27.0.0
raftos==0.2.6
simple-websocket==1.1.0
six==1.17.0
sqlalchemy==2.0.43
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing-extensions==4.13.2
tzdata==2025.2
wcwidth==0.2.13
werkzeug==3.0.6
wsproto==1.2.0
wtforms==3.1.2
zipp==3.20.2
